requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
boto3==1.26.30
botocore==1.29.30
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cytoolz==0.12.0
decorator==5.1.1
dnspython==2.2.1
eth-hash==0.5.1
eth-typing==3.2.0
eth-utils==2.1.0
fake-useragent==1.1.1
frozenlist==1.3.3
google-api-core==2.10.2
google-auth==2.14.1
google-cloud-bigquery==3.4.0
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.50.0
grpcio-status==1.50.0
html5lib==1.1
httptools==0.5.0
idna==3.4
jmespath==1.0.1
loguru==0.6.0
lxml==4.9.2
motor==3.1.1
msgpack==1.0.4
multidict==5.2.0
numpy==1.23.5
packaging==21.3
pandas==1.5.3
proto-plus==1.22.1
protobuf==4.21.9
py==1.11.0
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybase62==0.5.0
pybranca==0.5.0
pycoingecko==3.1.0
pycryptodome==3.16.0
pydantic==1.9.1
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
redis==4.3.5
requests==2.28.1
retry==0.9.2
rsa==4.9
s3transfer==0.6.0
sanic==21.6.0
Sanic-Cors==1.0.0
sanic-plugin-toolkit==1.2.1
sanic-routing==0.7.0
shroomdk==1.0.2
six==1.16.0
soupsieve==2.3.2.post1
toolz==0.12.0
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.11
uvloop==0.17.0
webencodings==0.5.1
websockets==10.4
yarl==1.8.1
